backend/requirements-prod.txt
# ViralClip Production Dependencies
# Optimized for Railway/Cloud deployment

# Core Framework
fastapi==0.128.0
uvicorn[standard]==0.40.0
python-multipart>=0.0.22
python-dotenv>=1.2.0
pydantic-settings==2.12.0
# Ensure Pydantic v2 compatibility
pydantic>=2.0.0

# Video Download
yt-dlp==2026.1.29

# AI - Transcription
# faster-whisper 1.x works best with numpy < 2.0
faster-whisper==1.2.1

# AI - Gemini (Viral Detection)
google-genai==0.2.1

# Video Processing - Scene Detection
# scenedetect requires opencv and numpy
scenedetect[opencv-headless]==0.6.5
# Fix: Pin numpy < 2.0 to avoid ABI issues with older precompiled wheels (common error)
numpy<2.0.0

# Cloud Storage
boto3==1.42.39

# Voice Dubbing
elevenlabs==2.33.1

# Async & WebSocket
websockets>=13.0,<15.0
aiofiles==25.1.0
aiosqlite==0.22.1
aiohttp>=3.13.0

# System
psutil>=7.2.0
